2019 - Day 09

+++ src/day05.rs
@@ -1,13 +1,13 @@
 use intcode_computer::{IntCodeComputer, OpCode};
 
-pub fn run_a(input: &Vec<String>) -> Vec<i32>{
+pub fn run_a(input: &Vec<String>) -> Vec<i64>{
     let unsupported = vec![OpCode::JumpIfTrue, OpCode::JumpIfFalse, OpCode::LessThan, OpCode::Equals];
     let computer = IntCodeComputer::new(unsupported);
 
     computer.run(&input, vec![1])
 }
 
-pub fn run_b(input: &Vec<String>) -> Vec<i32> {
+pub fn run_b(input: &Vec<String>) -> Vec<i64> {
     let computer = IntCodeComputer::new(vec![]);
 
     computer.run(&input, vec![5])
